+++ src/main/java/ru/nyansus/mc/protected_items/ProtectedItems.java
@@ -2,7 +2,7 @@ package ru.nyansus.mc.protected_items;
 
 import org.bukkit.plugin.java.JavaPlugin;
 
-public class IndestructibleItems extends JavaPlugin {
+public class ProtectedItems extends JavaPlugin {
 
     private Messages messages;
     private boolean protectionEnabled = true;
@@ -10,13 +10,15 @@ public class IndestructibleItems extends JavaPlugin {
     @Override
     public void onEnable() {
         saveDefaultConfig();
-        IndestructibleUtil.init(this);
+        ProtectionUtil.init(this);
         messages = new Messages(this);
-        IndestructibleCommand cmd = new IndestructibleCommand(this);
+        ProtectedCommand cmd = new ProtectedCommand(this);
         getCommand("protected").setExecutor(cmd);
         getCommand("protected").setTabCompleter(cmd);
-        getServer().getPluginManager().registerEvents(new IndestructibleListeners(this), this);
-        getServer().getHelpMap().addTopic(new IndestructibleHelpTopic(this));
+        getServer().getPluginManager().registerEvents(new DeathProtectionListener(this), this);
+        getServer().getPluginManager().registerEvents(new InventoryProtectionListener(this), this);
+        getServer().getPluginManager().registerEvents(new InteractionProtectionListener(this), this);
+        getServer().getHelpMap().addTopic(new ProtectedHelpTopic(this));
     }
 
     public Messages getMessages() {