+++ src/main/java/ru/nyansus/mc/protected_items/command/AddCommand.java
@@ -3,20 +3,21 @@ package ru.nyansus.mc.protected_items.command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
-import ru.nyansus.mc.protected_items.IndestructibleItems;
-import ru.nyansus.mc.protected_items.IndestructibleUtil;
+import ru.nyansus.mc.protected_items.ProtectedItems;
+import ru.nyansus.mc.protected_items.ProtectionUtil;
 
 import java.util.List;
 
 public final class AddCommand extends HeldItemCommand {
 
-    public AddCommand(IndestructibleItems plugin) {
+    public AddCommand(ProtectedItems plugin) {
         super(plugin);
     }
 
     @Override
     protected void executeWithItem(Player player, ItemStack hand, String[] args) {
-        IndestructibleUtil.setIndestructible(hand, true);
+        String loreText = plugin.getMessages().get(player, "item.protected-lore");
+        ProtectionUtil.setProtected(hand, true, loreText);
         player.sendMessage(plugin.getMessages().get(player, "command.add-success"));
     }